backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Redis Cache
redis==5.0.4
hiredis==2.3.2

# HTTP Client
httpx==0.27.0
aiohttp==3.9.5

# Web Scraping
newspaper3k==0.2.8
beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.31.0

# OCR
pytesseract==0.3.10
Pillow==10.3.0

# ML / NLP
transformers==4.41.1
torch==2.3.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2

# Validation & Settings
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Rate Limiting
slowapi==0.1.9

# Logging
loguru==0.7.2
